Lazy instantiate of KintoneApiService to delay setup HttpClient
