+++ infra/src/main/java/dev/mkeeda/spaceship/infra/api/KintoneApiService.kt
@@ -2,7 +2,6 @@ package dev.mkeeda.spaceship.infra.api
 
 import android.util.Base64
 import dev.mkeeda.spaceship.data.credential.LoginCredential
-import dev.mkeeda.spaceship.infra.datasource.LoginCredentialDataSource
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.header
@@ -10,12 +9,11 @@ import io.ktor.client.request.post
 import io.ktor.client.request.setBody
 import io.ktor.http.ContentType
 import io.ktor.http.contentType
-import javax.inject.Inject
 import kotlinx.serialization.Serializable
 
-internal class KintoneApiService @Inject constructor(
+internal class KintoneApiService(
     private val httpClient: HttpClient,
-    private val loginCredentialDataSource: LoginCredentialDataSource
+    private val loginCredential: LoginCredential
 ) {
     private val LoginCredential.authentication: String
         get() = Base64.encodeToString(
@@ -27,8 +25,6 @@ internal class KintoneApiService @Inject constructor(
         endpoint: KintoneApiEndpoint,
         param: KintoneApiEndpoint.RequestParam? = null
     ): T {
-        val loginCredential = loginCredentialDataSource.readLoginCredential()
-
         val response: SuccessResponse<T> = httpClient.post(loginCredential.domain + endpoint.path) {
             header("X-Cybozu-Authorization", loginCredential.authentication)
             contentType(ContentType.Application.Json)
